+++ src/schema/resolvers/topicResolvers.js
@@ -1,6 +1,6 @@
 export default {
   Topic: {
-    conferenceTopic: async ({ id }, data, { models: { ConferenceTopic } }) => {
+    conferenceTopics: async ({ id }, data, { models: { ConferenceTopic } }) => {
       const conferenceTopic = await ConferenceTopic.query().where(
         'topic_id',
         id,
@@ -79,9 +79,15 @@ export default {
     ) => {
       try {
         // delete ConferenceTopic with topic_id
-        await ConferenceTopic.query()
-          .delete()
-          .where('topic_id', id);
+        const confTopicWithTopicId = ConferenceTopic.query().where(
+          'topic_id',
+          id,
+        );
+        if (confTopicWithTopicId) {
+          await ConferenceTopic.query()
+            .delete()
+            .where('topic_id', id);
+        }
         const topic = await Topic.query().findById(id);
         await Topic.query().deleteById(id);
         return topic;
